requirements.txt
# MCPDESK Requirements
# Install: pip install -r requirements.txt

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0
nicegui>=1.4.0

# HTTP & API
requests>=2.31.0
python-multipart>=0.0.6

# AI & ML
google-genai>=0.3.0
mcp>=1.0.0
openai-whisper>=20231117

# Data & Finance
pandas>=2.0.0
yfinance>=0.2.28
scikit-learn>=1.3.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0

# Audio Processing (optional - for local voice features)
pyttsx3>=2.90
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.21.0

# Windows-specific audio dependencies
pywin32>=306; platform_system == "Windows"
